HBX-1910: Refactor JdbcMetadataBuilder into smaller responsibilities - Move class PropertyBinder from org.hibernate.tool.internal.reveng to org.hibernate.tool.internal.reveng.binder
