+++ orm/src/main/java/org/hibernate/tool/internal/reveng/binder/PropertyBinder.java
@@ -1,9 +1,10 @@
-package org.hibernate.tool.internal.reveng;
+package org.hibernate.tool.internal.reveng.binder;
 
 import org.hibernate.mapping.Property;
 import org.hibernate.mapping.Table;
 import org.hibernate.mapping.Value;
 import org.hibernate.tool.api.reveng.ReverseEngineeringStrategy;
+import org.hibernate.tool.internal.reveng.MetaAttributesBinder;
 import org.jboss.logging.Logger;
 
 public class PropertyBinder {
